+++ src/index.ts
@@ -1,3 +1,6 @@
+import { serveStdio } from "@modelcontextprotocol/server/stdio";
+
+import { createRedmineClientFromEnv } from "./redmine/client.js";
 import { createServer } from "./server.js";
 
-createServer();
+void serveStdio(() => createServer(createRedmineClientFromEnv()));